requirements.txt
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==3.2
django-modeltranslation==0.18.11
django-on-heroku==1.1.2
django-rosetta==0.9.9
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
fqdn==1.5.1
googletrans==4.0.0rc1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
isoduration==20.11.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
polib==1.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.15.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
stripe==2.56.0
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.6
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
whitenoise==6.6.0
zipp==3.17.0
